Fix deployment crash
I restored the apps/Gemfile because it is needed by .buildpacks. The gems first_app and second_app cannot be required by the root Gemfile because that would make the heroku-buildpack-ruby launch Rack too soon. Before that we need to install npm dependencies for the gems and build the ember-cli applications.